% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_db_pool.R
\name{mod_db_pool_teardown}
\alias{mod_db_pool_teardown}
\title{mod_db_pool_teardown: End Pool Connection}
\usage{
mod_db_pool_teardown(pool)
}
\arguments{
\item{pool}{Objeto pool::Pool}
}
\description{
mod_db_pool_teardown: End Pool Connection
}
